Organization Change
Changed the folder where I put code. I'm starting to want a lot of
methods that aren't directly asked for in HW, but useful for it. I'm
also finding it hard to sift through what code does what. Moved code
folder out of the individual class folders, sorted by language, and
added a zeroes folder in the MATLAB section.