Cactus Code Update
Finished the beginning of the cactus comma separated value file (csv)
reader. So far it reads a file into memory and determines the max and
min value of each column. Next I will have to make something to set each
row as a point in an appropriately sized grid function.