readCSV Min/Max
readCSV now works in cactus to load the entries in a comma separated
value file into memory and determine the min and max for each column in
the process. There is no grid variable support yet, so instead it
displays the min and max for each column which is being used to make the
grid variables adaptive.